Introduce .SQLLike operator
The default behavior of SQL LIKE is counter-intuitive when
used w/ NULL columns (also not processable w/
QualifierEvaluation).

When rendering to SQL, regular .Like now coalesces column NULL
values to the empty string. At the expense of the database
potentially not being able to use indices. Hence the .SQLLike
which is a straight mapping to the DB.