Very basic support for mom files
The mom files are just key-value archived files,
which we load w/ NSKeyValueUnarchiver but rather
by directly parsing the plist.

Probably makes no sense :-) But when building the
mobile version of the framework, Xcode will
convert the xcdatamodel into a .mom for apparently
no good reason (maybe because they also use
XMLDocument on macOS and iOS doesn't have that?)

Also included is a prepared XMLDocument mini-parser,
though the actual callbacks are not hooked up
(kinda trivial to finish up).