Maintain entity sort order for adaptor row SQL
Adaptor rows are Swift dictionaries and hence do not have stable
ordering. This resulted in SQL INSERTs/UPDATEs w/ changing column
ordering.
While that works, it is
a) hard to read during debugging
b) may affect caching in various places, locally or on server
The new code may be slightly slower in SQL building, but maintains
the order of the entity.