Rework SQLExpression handling
Do not format SQL as part of binding resolution, that is BAD.
This was just grabbing the description.

This has a limited but better support now. The binding is resolved
and for some known types the values are passed down and can then
be processed by SQLExpression during the actual formatting.