`qualifierWith(bindings:` => `qualifierWithBindings(`
Also make the returned value non-optional.
If a binding is not resolved, it sticks there for now.
TBD: Is that actually desirable, or is the old behaviour intentional?
Like if the resolved qualifier still has bound variables, should it
be dropped from the fetch spec? That was probably the original
intention, but doesn't seem helpful the way we use them :-)
TBD